Add error codes
The main reason for this is that I want to log (but not ignore) unknown
form fields:

	err = formam.Decode(r.Form, dst)
	if fErr, ok := err.(*formam.Error); ok && fErr.Code() == formam.ErrCodeUnknownField {
		zlog.Error(err)
		err = nil
	}
	return err

There was no way to get this information previously, and with
`IgnoreUnknownKeys` I would never know if there was an unknown key.

IMHO the `IgnoreUnknownKeys` option could be deprecated with this, as
it's now redundant. But that's a backwards-incompatible change (perhaps
something to consider for a v2?)

While I was changing `newError()`,  I also changed it to pass the field
and path and build the string with it.

Note: `strings.Builder` is Go 1.10 and newer only. Previously formam
supported a lot of quite old Go versions, and I took the liberty of
changing that. I can change it to use string concatenation if you
prefer, this is just a bit faster.