Call UnmarshalText for every value if the name ends with []
I have the following problem: one of my forms has a setting with a
bitmask, this all works grand, but right now there is, as far as I can
find, no way to tell Formam to decode that since it always uses the
first value.

So with my form:

	<input type="checkbox" name="settings.collect[]" value="1"> One</label>
	<input type="checkbox" name="settings.collect[]" value="2"> Two</label>
	<input type="checkbox" name="settings.collect[]" value="4"> Two</label>

Formam will just set the first value that is selected.

With this change, if the name ends with "[]" it will call
UnmarshalText() on all values. I think it might make sense to just
always do this, but that wouldn't be backwards-compatible so this is
probably better.